refactor(file): improve File API and handle stdin, memory safety, and line reading
- Introduced `FileAccessType` enum to explicitly define read/write mode.
- `File_Open` now copies the filename, sets the access type, and validates input.
- Added `File_OpenStdin()` for safe handling of stdin as a `File` object.
- Updated `File_Close` to safely free `path` and avoid closing stdin.
- `File_ReadLine` now uses `bytes_read` from `getline` instead of buffer length.
- Added parameter validation to `File_ReadLine` and `File_WriteLine`.
- Updated UTF8String conversion to `const UTF8String*` in `UTF8String_ToStr`.

+++ src/common/utf8string.c
@@ -58,7 +58,7 @@ void UTF8String_Free(UTF8String *str) {
 
 char *UTF8String_ToStr(const UTF8String *string) {
     size_t size = 0;
-    for (int i=0; i<string->length; i++) {
+    for (size_t i=0; i<string->length; i++) {
         size += string->chars[i].length;
     }
     char *return_str = malloc(sizeof(char) * size + 1);
@@ -66,7 +66,7 @@ char *UTF8String_ToStr(const UTF8String *string) {
         logFatal("Cannot allocate memory for return string.");
     }
     size_t idx = 0;
-    for (int i=0; i<string->length; i++) {
+    for (size_t i=0; i<string->length; i++) {
         for (int j=0; j<string->chars[i].length; j++) {
             return_str[idx] = string->chars[i].bytes[j];
             idx++;
